2024/02/03 更新membership/controller/MembershipController 裡面的update_Status，我帳戶狀態直接用定時器，沒有在透過Redis。我前台header加上連結(首頁+faq+聯絡我們)。
