+++ src/main/java/com/membership/controller/LoginController.java
@@ -9,21 +9,30 @@ import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.ui.ModelMap;
 import org.springframework.util.DigestUtils;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import com.membership.model.MembershipVO;
 import com.membership.service.MailService;
 import com.membership.service.MembershipService;
+
 
 import com.membership.service.RedisService;
 
 @Controller
 @RequestMapping("/membership")
 public class LoginController {
+	
+	@GetMapping("ZuoHuo")
+	public String ZuoHuo(ModelMap model) {
+		return "front-end/zuo-huo";
+	}
 
+	
 	@Autowired
 	MembershipService membershipSvc;
