+++ src/main/resources/templates/front-end/membership/addMembership.html
@@ -171,13 +171,13 @@ h6#memPic {
 
 			<ul
 				class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
-				<li><a href="index.html" class="nav-link px-2 link-secondary">首頁</a></li>
+				<li><a th:href="@{/membership/ZuoHuo}" class="nav-link px-2 link-secondary">首頁</a></li>
 				<li><a href="join.html" class="nav-link px-2 link-secondary">瀏覽活動</a></li>
 				<li><a href="create.html" class="nav-link px-2 link-secondary">舉辦活動</a></li>
 				<li><a href="rent.html" class="nav-link px-2 link-secondary">場地租借</a></li>
-				<li><a href="customer.html"
+				<li><a th:href="@{/faq/faqs}" 
 					class="nav-link px-2 link-secondary">FAQs</a></li>
-				<li><a href="about.html" class="nav-link px-2 link-secondary">關於我們</a></li>
+				<li><a th:href="@{/faq/faqs}"  class="nav-link px-2 link-secondary">關於我們</a></li>
 			</ul>
 
 			<div class="col-md-3 text-end">
